Implemented configurable listening address
When applied, this commit tries to implement the feature requests reported on the issues #43 and #227.
Basically, with this, it is now possible to make the server listen only on certain addresses.
This increases security and allows the typical "SSH usage", that is, SSH + publickey auth + local port forwarding + VNC listening on localhost only.

To achieve this:

- on droidvnc-ng.c, vncServerStart was modified in order to provide a further "jstring listenIf" parameter.
It uses rfbScreenInfo*'s listenInterface property.

- A TableRow was added in order to allow users to input the desired address to use ("Listening Address").

- On droidvnc-ng.c, the method vncServerGetListenInterface was introduced in order to track if the server is currenctly set to listen to 0.0.0.0 or not (this is used to properly show what addresses are available to use on the UI).

Please note that, if an invalid address/host is set, by default, it is assumed the address 0.0.0.0 is requested.