MainService+Activity: Spinner for Interface listening selection
When applied, this commit should improve the previous one, in relation to the address binding.
As requested by @bk138, now, instead of an EditText, a Spinner with the available interfaces is used.
This is how it is achieved:

- Re-using some of the methods already made by @bk138, all the available up network interfaces which have AT LEAST ONE IPv4 address are collected.
These will be shown into the Spinner by using the class ListenIfAdapter;

- The first option of the Spinner is always "0.0.0.0", so that one can use the previous default behaviour (listen to any interface);

- Some earlier modifications are not needed anymore, so they were removed (e.g. vncGetListenInterface in cpp);